+++ src/main/java/io/github/slimeistdev/semaphore/utils/AuthUtils.java
@@ -18,22 +18,28 @@
 
 package io.github.slimeistdev.semaphore.utils;
 
+import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.minecraft.commands.CommandSourceStack;
-import net.minecraft.network.chat.Component;
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Items;
-import net.minecraft.world.level.Level;
 
-import java.util.List;
-import java.util.function.Supplier;
-
+@SuppressWarnings("unused")
 public class AuthUtils {
-    public static boolean isAuthorized(Player player) {
-        return player.hasPermissions(2);
+    public static final String DEBUG_PERMISSION = "semaphore.debug";
+    public static final String TELEPORT_PERMISSION = "semaphore.teleport";
+
+    public static boolean isDebugAuthorized(Player player) {
+        return Permissions.check(player, DEBUG_PERMISSION, 2);
     }
 
-    public static boolean isAuthorized(CommandSourceStack commandSourceStack) {
-        return commandSourceStack.hasPermission(2);
+    public static boolean isDebugAuthorized(CommandSourceStack commandSourceStack) {
+        return Permissions.check(commandSourceStack, DEBUG_PERMISSION, 2);
+    }
+
+    public static boolean isTeleportAuthorized(Player player) {
+        return Permissions.check(player, TELEPORT_PERMISSION, 2);
+    }
+
+    public static boolean isTeleportAuthorized(CommandSourceStack commandSourceStack) {
+        return Permissions.check(commandSourceStack, TELEPORT_PERMISSION, 2);
     }
 }